chore: add dash in menu button and split button name
