+++ packages/html/src/menu-button/index.js
@@ -1,0 +1,1 @@
+export * from './menu-button.jsx';
