--- packages/html/src/menubutton/index.js
@@ -1,1 +1,0 @@
-export * from './menubutton.jsx';
