+++ packages/html/src/split-button/index.js
@@ -1,0 +1,1 @@
+export * from './split-button.jsx';
