--- packages/html/src/splitbutton/index.js
@@ -1,1 +1,0 @@
-export * from './splitbutton.jsx';
